use opensshs control socket to reuse an already established connection
Add a new option -S (resembling the same as that from ssh) to use this
socket from an openssh control master for the connection. Saving two forks
and the authorization procedure.

When the socket name is "none" than that is passed to the ssh command, so
that it does not connect/create a control socket.

The files fdpass.* come directly from the openssh source package, I haven't
the skill to code the fd passing over an unix socket. If that would be a
licence problem, I think someone can code this again.

Signed-off-by: Bert Wesarg <bert.wesarg@googlemail.com>

---

Maybe we should fallback to the normal ssh command, when the control socket
does not work?

The subsystem is hardcoded to "sftp", I don't know if it makes sense to let
the user specify the subsystem.